requirements.txt
# Core dependencies
pysaml2
flask==2.2.5
flask-login==0.6.2
python-dotenv==1.0.0
pymongo==4.5.0
mongoengine==0.27.0
pyjwt==2.8.0
cryptography==41.0.4

# SAML dependencies
python3-saml==1.16.0

# OIDC dependencies
authlib==1.2.1
requests==2.31.0

# Development and testing
pytest==7.4.0
black==23.7.0
flake8==6.1.0

werkzeug<3.0
mongomock==4.1.2

signxml>=3.0.0

lxml>=4.9.0
